fix(apple): wrap AgX sample_lut return in explicit float() cast
`return sampler.sample(...).r` returns `half` from a `float`-declared
function. Modern macOS CoreImage's CIKernel.kernels(withMetalString:)
rejects the implicit half→float widening with a Metal compile error,
which fires the AgX kernel-load assertion at app launch.

Wrap with float(...) to match the explicit-cast convention introduced
in commit f720a0f (sample() float4 wrapping).

+++ src/apple/Packages/MapleCore/Sources/MapleCore/Metal/AgXViewTransform.metal
@@ -29,7 +29,7 @@ float agx_log_encode(float linear) {
 
 /// Sample the 1D LUT texture at normalized position t ∈ [0, 1].
 float sample_lut(coreimage::sampler_h lut_sampler, float t) {
-    return lut_sampler.sample(float2(t * (AGX_LUT_SIZE - 1.0) / AGX_LUT_SIZE, 0.0)).r;
+    return float(lut_sampler.sample(float2(t * (AGX_LUT_SIZE - 1.0) / AGX_LUT_SIZE, 0.0)).r);
 }
 
 /// Apply contrast modulation: expand/compress around MID_NORM
